backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
qdrant-client==1.7.0
sentence-transformers==2.3.1
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
tiktoken==0.5.2
python-dotenv==1.0.0
aiofiles==23.2.1


